refactor: update imports and parameter types for consistency across services
